requirements.txt
anthropic>=0.93.0
google-api-python-client>=2.194.0
google-auth-oauthlib>=1.3.1
requests>=2.33.0
